chore(deps): bump dependencies, cleanup project
updated project dependencies;
verified project workability after update;
removed junky
files;
revised and fixed e2e specs
;added e2e testing pipeline;

BREAKING CHANGE: reinstall project dependencies

+++ decorate-angular-cli.js
@@ -29,43 +29,8 @@ let output;
 try {
   output = require('@nrwl/workspace').output;
 } catch (e) {
-  console.warn(
-    'Angular CLI could not be decorated to enable computation caching. Please ensure @nrwl/workspace is installed.',
-  );
+  console.warn('Angular CLI could not be decorated to enable computation caching. Please ensure @nrwl/workspace is installed.');
   process.exit(0);
-}
-
-/**
- * Paths to files being patched
- */
-const angularCLIInitPath = 'node_modules/@angular/cli/lib/cli/index.js';
-
-/**
- * Patch index.js to warn you if you invoke the undecorated Angular CLI.
- */
-function patchAngularCLI(initPath) {
-  const angularCLIInit = fs.readFileSync(initPath, 'utf-8').toString();
-
-  if (!angularCLIInit.includes('NX_CLI_SET')) {
-    fs.writeFileSync(
-      initPath,
-      `
-if (!process.env['NX_CLI_SET']) {
-  const { output } = require('@nrwl/workspace');
-  output.warn({ title: 'The Angular CLI was invoked instead of the Nx CLI. Use "npx ng [command]" or "nx [command]" instead.' });
-}
-
-if (process.argv[2] === 'update') {
-  const { output } = require('@nrwl/workspace');
-  output.error({
-    title: '"ng update" is deprecated in favor of "nx migrate". Read more: https://nx.dev/latest/angular/workspace/update'
-  });
-  throw new Error();
-}
-${angularCLIInit}
-    `,
-    );
-  }
 }
 
 /**
@@ -82,25 +47,23 @@ function symlinkNgCLItoNxCLI() {
        * Such that it works in all shells and works with npx.
        */
       ['', '.cmd', '.ps1'].forEach(ext => {
-        if (fs.existsSync(nxPath + ext))
-          fs.writeFileSync(ngPath + ext, fs.readFileSync(nxPath + ext));
+        if (fs.existsSync(nxPath + ext)) fs.writeFileSync(ngPath + ext, fs.readFileSync(nxPath + ext));
       });
     } else {
       // If unix-based, symlink
       cp.execSync(`ln -sf ./nx ${ngPath}`);
     }
-  } catch (e) {
-    output.error({
-      title: 'Unable to create a symlink from the Angular CLI to the Nx CLI:' + e.message,
-    });
+  }
+  catch(e) {
+    output.error({ title: 'Unable to create a symlink from the Angular CLI to the Nx CLI:' + e.message });
     throw e;
   }
 }
 
 try {
   symlinkNgCLItoNxCLI();
-  patchAngularCLI(angularCLIInitPath);
+  require('@nrwl/cli/lib/decorate-cli').decorateCli();
   output.log({ title: 'Angular CLI has been decorated to enable computation caching.' });
-} catch (e) {
+} catch(e) {
   output.error({ title: 'Decoration of the Angular CLI did not complete successfully' });
 }
